@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.15.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -30,20 +30,22 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  camera:
-  image_picker:
-  provider:
-  hive:
-  hive_flutter:
-  path_provider:
-  permission_handler:
-  contacts_service: ^0.4.6
-  app_settings: ^4.0.4
-  flutter_svg: ^0.19.0
-  badges: ^1.1.1
+  app_settings: 4.1.1
+  badges: 2.0.2
+  camera: 0.9.4+8
+  collection: 1.15.0
+  contacts_service: 0.6.3
+  cupertino_icons: ^1.0.4
+  flutter_svg: 1.0.3
+  hive: 2.0.5
+  hive_flutter: 1.1.0
+  image_picker: 0.8.4+5
+  path_provider: 2.0.8
+  permission_handler: 8.3.0
+  provider: 6.0.2
 
 dev_dependencies:
+  flutter_lints: ^1.0.4
   flutter_test:
     sdk: flutter
 
