@@ -1,9 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:whatsapp_template/app_services/call_notifier.dart';
-import 'package:whatsapp_template/app_utils/app_theme.dart';
-import 'package:whatsapp_template/app_utils/ui_components.dart';
+
+import '../../../app_services/call_notifier.dart';
+import '../../../app_utils/app_theme.dart';
+import '../../../app_utils/ui_components.dart';
 
 class CallTab extends StatefulWidget {
+  const CallTab({Key? key}) : super(key: key);
+
   @override
   _CallTabState createState() => _CallTabState();
 }
@@ -17,18 +20,18 @@ class _CallTabState extends State<CallTab> {
           switch (snapshot.connectionState) {
             case ConnectionState.done:
               if (snapshot.hasData) {
-                List<Map<String, dynamic>> _callData = [];
-                _callData = snapshot.data;
+                List<Map<String, dynamic>>? _callData = [];
+                _callData = snapshot.data as List<Map<String, dynamic>>?;
                 return ListView.separated(
                   physics: BouncingScrollPhysics(),
                   key: PageStorageKey('callTabKey'),
                   shrinkWrap: true,
-                  itemCount: _callData.length,
+                  itemCount: _callData!.length,
                   itemBuilder: (context, index) {
                     return BuildCallListTile(
                       context: context,
                       index: index,
-                      callData: _callData[index],
+                      callData: _callData![index],
                     );
                   },
                   separatorBuilder: (context, index) {
@@ -41,9 +44,8 @@ class _CallTabState extends State<CallTab> {
                   },
                 );
               } else {
-                return NoRecordsFound("No Call Records");
+                return NoRecordsFound(message: 'No Call Records');
               }
-              break;
             case ConnectionState.waiting:
               return Padding(
                 padding: const EdgeInsets.all(10.0),
@@ -53,10 +55,8 @@ class _CallTabState extends State<CallTab> {
                   ),
                 ),
               );
-              break;
             default:
               return Container();
-              break;
           }
         });
   }
