@@ -1,17 +1,35 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:whatsapp_template/app_services/theme_notifier.dart';
-import 'package:whatsapp_template/app_utils/hive_config.dart';
-import 'package:whatsapp_template/app_utils/size_config.dart';
-import 'package:whatsapp_template/app_utils/ui_components.dart';
-import 'package:whatsapp_template/app_utils/util_functions.dart';
+
+import '../../app_services/theme_notifier.dart';
+import '../../app_utils/hive_config.dart';
+import '../../app_utils/size_config.dart';
+import '../../app_utils/ui_components.dart';
+import '../../app_utils/util_functions.dart';
 
 class SettingsChatsScreen extends StatefulWidget {
+  const SettingsChatsScreen({Key? key}) : super(key: key);
+
   @override
   _SettingsChatsScreenState createState() => _SettingsChatsScreenState();
 }
 
 class _SettingsChatsScreenState extends State<SettingsChatsScreen> {
+  ValueNotifier<int>? _themeGroupValueNotifier;
+  ValueNotifier<int>? _fontSizeGroupValueNotifier;
+  ValueNotifier<bool>? _switch1Notifier;
+  ValueNotifier<bool>? _switch2Notifier;
+  final List<String> _themeList = <String>[
+    'System default',
+    'Light',
+    'Dark',
+  ];
+  final List<String> _fontSizeList = <String>[
+    'Small',
+    'Medium',
+    'Large',
+  ];
+
   @override
   void initState() {
     super.initState();
@@ -24,36 +42,12 @@ class _SettingsChatsScreenState extends State<SettingsChatsScreen> {
   }
 
   @override
-  void dispose() {
-    _themeGroupValueNotifier.dispose();
-    _fontSizeGroupValueNotifier.dispose();
-    _switch1Notifier.dispose();
-    _switch2Notifier.dispose();
-    super.dispose();
-  }
-
-  ValueNotifier<int> _themeGroupValueNotifier;
-  ValueNotifier<int> _fontSizeGroupValueNotifier;
-  ValueNotifier<bool> _switch1Notifier;
-  ValueNotifier<bool> _switch2Notifier;
-  List<String> _themeList = <String>[
-    "System default",
-    "Light",
-    "Dark",
-  ];
-  List<String> _fontSizeList = <String>[
-    "Small",
-    "Medium",
-    "Large",
-  ];
-
-  @override
   Widget build(BuildContext context) {
-    SizeConfig()..init(context);
+    SizeConfig().init(context);
     return SafeArea(
       child: Scaffold(
         appBar: AppBar(
-          title: Text("Chats"),
+          title: Text('Chats'),
           leading: IconButton(
             icon: Icon(Icons.arrow_back),
             iconSize: SizeConfig.textScaleFactor * 25,
@@ -69,11 +63,11 @@ class _SettingsChatsScreenState extends State<SettingsChatsScreen> {
             crossAxisAlignment: CrossAxisAlignment.start,
             mainAxisSize: MainAxisSize.min,
             children: [
-              BuildHeading(context: context, title: "Display"),
+              BuildHeading(context: context, title: 'Display'),
               BuildSettingsListTile(
                   leadingImage: 'assets/icons/brightness.svg',
                   context: context,
-                  title: "Theme",
+                  title: 'Theme',
                   subtitle: _themeList[
                       int.parse(HiveConfig.hiveReadData('ThemeMode'))],
                   callback: () async {
@@ -81,14 +75,14 @@ class _SettingsChatsScreenState extends State<SettingsChatsScreen> {
                       context: context,
                       groupValueNotifier: _themeGroupValueNotifier,
                     ).then((res) {
-                      if (res != null && res.toString() == "OK") {
+                      if (res != null && res.toString() == 'OK') {
                         HiveConfig.hiveCreateData('ThemeMode',
-                            _themeGroupValueNotifier.value.toString());
+                            _themeGroupValueNotifier!.value.toString());
                         Provider.of<ThemeNotifier>(context, listen: false)
                             .setThemeMode(getThemeModeByIndex(
-                                _themeGroupValueNotifier.value));
+                                _themeGroupValueNotifier!.value));
                       } else {
-                        _themeGroupValueNotifier.value =
+                        _themeGroupValueNotifier!.value =
                             int.parse(HiveConfig.hiveReadData('ThemeMode'));
                       }
                     });
@@ -96,38 +90,38 @@ class _SettingsChatsScreenState extends State<SettingsChatsScreen> {
               BuildSettingsListTile(
                 context: context,
                 leadingIcon: Icons.wallpaper,
-                title: "Wallpaper",
+                title: 'Wallpaper',
               ),
               Divider(thickness: 0.6),
-              BuildHeading(context: context, title: "Chat Settings"),
+              BuildHeading(context: context, title: 'Chat Settings'),
               BuildSettingsListTile(
                 context: context,
-                title: "Enter is send",
+                title: 'Enter is send',
                 contentPadding: EdgeInsets.only(left: 70.0),
-                subtitle: "Enter key will send your message",
+                subtitle: 'Enter key will send your message',
                 trailing: BuildSwitch(
                     valueNotifier: _switch1Notifier,
                     callback: (value) {
-                      _switch1Notifier.value = value;
+                      _switch1Notifier!.value = value;
                     }),
               ),
               BuildSettingsListTile(
                 context: context,
-                title: "Media visibility",
+                title: 'Media visibility',
                 contentPadding: EdgeInsets.only(left: 70.0),
                 subtitle: "Show newly downloaded media in your phone's gallery",
                 trailing: BuildSwitch(
                     valueNotifier: _switch2Notifier,
                     callback: (value) {
-                      _switch2Notifier.value = value;
+                      _switch2Notifier!.value = value;
                     }),
               ),
               ValueListenableBuilder(
-                valueListenable: _fontSizeGroupValueNotifier,
-                builder: (context, value, _) {
+                valueListenable: _fontSizeGroupValueNotifier!,
+                builder: (context, dynamic value, _) {
                   return BuildSettingsListTile(
                       context: context,
-                      title: "Font Size",
+                      title: 'Font Size',
                       contentPadding: EdgeInsets.only(left: 70.0),
                       subtitle: _fontSizeList[
                           int.parse(HiveConfig.hiveReadData('FontSize'))],
@@ -136,44 +130,53 @@ class _SettingsChatsScreenState extends State<SettingsChatsScreen> {
                           context: context,
                           groupValueNotifier: _fontSizeGroupValueNotifier,
                           callback1: (val) {
-                            _fontSizeGroupValueNotifier.value = val;
+                            _fontSizeGroupValueNotifier!.value = val;
                             Navigator.pop(context);
                           },
                           callback2: (val) {
-                            _fontSizeGroupValueNotifier.value = val;
+                            _fontSizeGroupValueNotifier!.value = val;
                             Navigator.pop(context);
                           },
                           callback3: (val) {
-                            _fontSizeGroupValueNotifier.value = val;
+                            _fontSizeGroupValueNotifier!.value = val;
                             Navigator.pop(context);
                           },
                         ).then((_) {
                           HiveConfig.hiveCreateData('FontSize',
-                              '${_fontSizeGroupValueNotifier.value}');
+                              '${_fontSizeGroupValueNotifier!.value}');
                         });
                       });
                 },
               ),
               BuildSettingsListTile(
                 context: context,
-                title: "App Language",
+                title: 'App Language',
                 contentPadding: EdgeInsets.only(left: 70.0),
                 subtitle: "Phone's language (English)",
               ),
               BuildSettingsListTile(
                 context: context,
                 leadingIcon: Icons.backup,
-                title: "Chat backup",
+                title: 'Chat backup',
               ),
               BuildSettingsListTile(
                 context: context,
                 leadingIcon: Icons.history,
-                title: "Chat history",
+                title: 'Chat history',
               ),
             ],
           ),
         ),
       ),
     );
+  }
+
+  @override
+  void dispose() {
+    _themeGroupValueNotifier!.dispose();
+    _fontSizeGroupValueNotifier!.dispose();
+    _switch1Notifier!.dispose();
+    _switch2Notifier!.dispose();
+    super.dispose();
   }
 }
