@@ -1,166 +1,166 @@
-import 'package:whatsapp_template/app_utils/util_functions.dart';
+import '../app_utils/util_functions.dart';
 
 class CallNotifier {
-  static List<Map<String, dynamic>> _callData = [
+  static final List<Map<String, dynamic>> _callData = [
     {
-      "contactName": "Flutter Developer",
-      "contactProfilePic": "assets/profiles/profile6.jpeg",
-      "callStatus": CallStatus.call_missed_incoming,
-      "callFrequency": "2",
-      "callTime": "September 14, 3:38 PM",
-      "callType": CallType.voice,
+      'contactName': 'Flutter Developer',
+      'contactProfilePic': 'assets/profiles/profile6.jpeg',
+      'callStatus': CallStatus.call_missed_incoming,
+      'callFrequency': '2',
+      'callTime': 'September 14, 3:38 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "John",
-      "contactProfilePic": "assets/profiles/profile2.jpeg",
-      "callStatus": CallStatus.call_made,
-      "callFrequency": "1",
-      "callTime": "September 3, 4:56 PM",
-      "callType": CallType.voice,
+      'contactName': 'John',
+      'contactProfilePic': 'assets/profiles/profile2.jpeg',
+      'callStatus': CallStatus.call_made,
+      'callFrequency': '1',
+      'callTime': 'September 3, 4:56 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Kajal",
-      "contactProfilePic": "assets/profiles/profile1.jpg",
-      "callStatus": CallStatus.call_received,
-      "callFrequency": "5",
-      "callTime": "September 3, 4:54 PM",
-      "callType": CallType.video,
+      'contactName': 'Kajal',
+      'contactProfilePic': 'assets/profiles/profile1.jpg',
+      'callStatus': CallStatus.call_received,
+      'callFrequency': '5',
+      'callTime': 'September 3, 4:54 PM',
+      'callType': CallType.video,
     },
     {
-      "contactName": "Sanju",
-      "contactProfilePic": "assets/profiles/profile3.jpeg",
-      "callStatus": CallStatus.call_received,
-      "callFrequency": "2",
-      "callTime": "August 29, 3:10 PM",
-      "callType": CallType.voice,
+      'contactName': 'Sanju',
+      'contactProfilePic': 'assets/profiles/profile3.jpeg',
+      'callStatus': CallStatus.call_received,
+      'callFrequency': '2',
+      'callTime': 'August 29, 3:10 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Albert",
-      "contactProfilePic": "assets/profiles/profile5.jpeg",
-      "callStatus": CallStatus.call_missed_outgoing,
-      "callFrequency": "1",
-      "callTime": "August 29, 10:33 AM",
-      "callType": CallType.voice,
+      'contactName': 'Albert',
+      'contactProfilePic': 'assets/profiles/profile5.jpeg',
+      'callStatus': CallStatus.call_missed_outgoing,
+      'callFrequency': '1',
+      'callTime': 'August 29, 10:33 AM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Arjun",
-      "contactProfilePic": "",
-      "callStatus": CallStatus.call_missed_incoming,
-      "callFrequency": "1",
-      "callTime": "August 26, 10:43 AM",
-      "callType": CallType.voice,
+      'contactName': 'Arjun',
+      'contactProfilePic': '',
+      'callStatus': CallStatus.call_missed_incoming,
+      'callFrequency': '1',
+      'callTime': 'August 26, 10:43 AM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Keerthi",
-      "contactProfilePic": "assets/profiles/profile4.jpeg",
-      "callStatus": CallStatus.call_received,
-      "callFrequency": "1",
-      "callTime": "August 25, 6:09 PM",
-      "callType": CallType.voice,
+      'contactName': 'Keerthi',
+      'contactProfilePic': 'assets/profiles/profile4.jpeg',
+      'callStatus': CallStatus.call_received,
+      'callFrequency': '1',
+      'callTime': 'August 25, 6:09 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Kajal",
-      "contactProfilePic": "assets/profiles/profile1.jpg",
-      "callStatus": CallStatus.call_made,
-      "callFrequency": "5",
-      "callTime": "August 25, 10:25 AM",
-      "callType": CallType.video,
+      'contactName': 'Kajal',
+      'contactProfilePic': 'assets/profiles/profile1.jpg',
+      'callStatus': CallStatus.call_made,
+      'callFrequency': '5',
+      'callTime': 'August 25, 10:25 AM',
+      'callType': CallType.video,
     },
     {
-      "contactName": "John",
-      "contactProfilePic": "assets/profiles/profile2.jpeg",
-      "callStatus": CallStatus.call_made,
-      "callFrequency": "2",
-      "callTime": "August 24, 7:59 PM",
-      "callType": CallType.voice,
+      'contactName': 'John',
+      'contactProfilePic': 'assets/profiles/profile2.jpeg',
+      'callStatus': CallStatus.call_made,
+      'callFrequency': '2',
+      'callTime': 'August 24, 7:59 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Sanju",
-      "contactProfilePic": "assets/profiles/profile3.jpeg",
-      "callStatus": CallStatus.call_missed_incoming,
-      "callFrequency": "1",
-      "callTime": "August 24, 7:48 PM",
-      "callType": CallType.voice,
+      'contactName': 'Sanju',
+      'contactProfilePic': 'assets/profiles/profile3.jpeg',
+      'callStatus': CallStatus.call_missed_incoming,
+      'callFrequency': '1',
+      'callTime': 'August 24, 7:48 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Albert",
-      "contactProfilePic": "assets/profiles/profile5.jpeg",
-      "callStatus": CallStatus.call_received,
-      "callFrequency": "3",
-      "callTime": "August 24, 7:07 PM",
-      "callType": CallType.voice,
+      'contactName': 'Albert',
+      'contactProfilePic': 'assets/profiles/profile5.jpeg',
+      'callStatus': CallStatus.call_received,
+      'callFrequency': '3',
+      'callTime': 'August 24, 7:07 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Keerthi",
-      "contactProfilePic": "assets/profiles/profile4.jpeg",
-      "callStatus": CallStatus.call_received,
-      "callFrequency": "1",
-      "callTime": "August 24, 5:32 PM",
-      "callType": CallType.video,
+      'contactName': 'Keerthi',
+      'contactProfilePic': 'assets/profiles/profile4.jpeg',
+      'callStatus': CallStatus.call_received,
+      'callFrequency': '1',
+      'callTime': 'August 24, 5:32 PM',
+      'callType': CallType.video,
     },
     {
-      "contactName": "Kajal",
-      "contactProfilePic": "assets/profiles/profile1.jpg",
-      "callStatus": CallStatus.call_received,
-      "callFrequency": "5",
-      "callTime": "August 24, 5:20 PM",
-      "callType": CallType.voice,
+      'contactName': 'Kajal',
+      'contactProfilePic': 'assets/profiles/profile1.jpg',
+      'callStatus': CallStatus.call_received,
+      'callFrequency': '5',
+      'callTime': 'August 24, 5:20 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Arjun",
-      "contactProfilePic": "",
-      "callStatus": CallStatus.call_split,
-      "callFrequency": "1",
-      "callTime": "August 24, 5:07 PM",
-      "callType": CallType.voice,
+      'contactName': 'Arjun',
+      'contactProfilePic': '',
+      'callStatus': CallStatus.call_split,
+      'callFrequency': '1',
+      'callTime': 'August 24, 5:07 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "John",
-      "contactProfilePic": "assets/profiles/profile2.jpeg",
-      "callStatus": CallStatus.call_made,
-      "callFrequency": "1",
-      "callTime": "August 24, 4:04 PM",
-      "callType": CallType.voice,
+      'contactName': 'John',
+      'contactProfilePic': 'assets/profiles/profile2.jpeg',
+      'callStatus': CallStatus.call_made,
+      'callFrequency': '1',
+      'callTime': 'August 24, 4:04 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Keerthi",
-      "contactProfilePic": "assets/profiles/profile4.jpeg",
-      "callStatus": CallStatus.call_merge,
-      "callFrequency": "2",
-      "callTime": "August 24, 2:12 PM",
-      "callType": CallType.voice,
+      'contactName': 'Keerthi',
+      'contactProfilePic': 'assets/profiles/profile4.jpeg',
+      'callStatus': CallStatus.call_merge,
+      'callFrequency': '2',
+      'callTime': 'August 24, 2:12 PM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Flutter Developer",
-      "contactProfilePic": "assets/profiles/profile6.jpeg",
-      "callStatus": CallStatus.call_made,
-      "callFrequency": "3",
-      "callTime": "August 24, 10:57 AM",
-      "callType": CallType.video,
+      'contactName': 'Flutter Developer',
+      'contactProfilePic': 'assets/profiles/profile6.jpeg',
+      'callStatus': CallStatus.call_made,
+      'callFrequency': '3',
+      'callTime': 'August 24, 10:57 AM',
+      'callType': CallType.video,
     },
     {
-      "contactName": "Kajal",
-      "contactProfilePic": "assets/profiles/profile1.jpg",
-      "callStatus": CallStatus.call_missed_incoming,
-      "callFrequency": "1",
-      "callTime": "August 24, 10:21 AM",
-      "callType": CallType.voice,
+      'contactName': 'Kajal',
+      'contactProfilePic': 'assets/profiles/profile1.jpg',
+      'callStatus': CallStatus.call_missed_incoming,
+      'callFrequency': '1',
+      'callTime': 'August 24, 10:21 AM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Albert",
-      "contactProfilePic": "assets/profiles/profile5.jpeg",
-      "callStatus": CallStatus.call_made,
-      "callFrequency": "3",
-      "callTime": "August 24, 10:16 AM",
-      "callType": CallType.voice,
+      'contactName': 'Albert',
+      'contactProfilePic': 'assets/profiles/profile5.jpeg',
+      'callStatus': CallStatus.call_made,
+      'callFrequency': '3',
+      'callTime': 'August 24, 10:16 AM',
+      'callType': CallType.voice,
     },
     {
-      "contactName": "Keerthi",
-      "contactProfilePic": "assets/profiles/profile4.jpeg",
-      "callStatus": CallStatus.call_split,
-      "callFrequency": "1",
-      "callTime": "August 21, 10:56 PM",
-      "callType": CallType.voice,
+      'contactName': 'Keerthi',
+      'contactProfilePic': 'assets/profiles/profile4.jpeg',
+      'callStatus': CallStatus.call_split,
+      'callFrequency': '1',
+      'callTime': 'August 21, 10:56 PM',
+      'callType': CallType.voice,
     },
   ];
 
