@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 
 extension ColorExtension on String {
   Color toHexColor() {
-    var hexColor = this.replaceAll("#", "");
+    var hexColor = replaceAll('#', '');
     if (hexColor.length == 6) {
-      hexColor = "FF" + hexColor;
+      hexColor = 'FF' + hexColor;
     }
-    return Color(int.parse("0x$hexColor"));
+    return Color(int.parse('0x$hexColor'));
   }
 }
